Create a Platform script and add it to the platform prefabs
First we’ll create a Scripts folder and then in it we’ll create a Platform.cs script and attach it to each of the four Platform prefabs. In the script we’ll create a Mobility enum and add the basic variables. We’ll access the Rigidbody2D component in the Start method and we’ll handle the movement pattern of mobile platforms (from right to left to right, etc.) in the FixedUpdate method.