Add a Player GameObject and implement horizontal movement and basic jumping
We’ll add a Player GameObject to the scene and attach a BoxCollider2D and a Rigidbody2D to it. Then we’ll create a PlayerController script and attach it to the Player. In the script we’ll implement horizontal movement and basic jumping. Finally, we’ll turn the Player into a prefab.