Add a ground check to the player and implement double-jumping
We’ll create an empty object as a child of the Player and name it GroundCheck. We’ll add a GroundCheck script to it. Its task is to check if the player is standing on a platform. If the player is grounded on a platform, they can double-jump. Otherwise, if they are mid-air, they can’t.