Implement the Kick coroutine to move a platform when the Player lands on it.
First we’ll add the “Player” tag to the Player prefab. Then we’ll check for collisions between a platform and the Player. Whenever the Player collides with a platform from above (so lands on it instead of hitting it with their head from below), they become the child of the platform. If the platform is a MobileOnEnter type, a coroutine is started to move the platform. When the Player leaves the platform, it stops moving.