Create the Forest prefab with a forest background and a floor trigger collider to restart the scene
In each level there will be a different background image of the forest. If the player falls off a platform and falls down to the forest floor, they will die and the scene will be restarted. We’ll create a Forest prefab that contains the background image and the floor trigger collider. Then we’ll add the prefab to each scene.