Use the Cinemachine package to make the camera follow the Player
After installing the Cinemachine package we select the Player as the object to follow and we add a camera confiner around the background image so that the camera doesn’t see anything beyond the forest. We create the Camera Setup prefab that contains the main camera, the virtual camera delivered by the package and the camera confiner. Then we replace the main camera in each scene by an instance of the Camera Setup prefab.