Add the OnTriggerEnter2D method to PlayerController to check for collisions with props
The Player will sometimes bump into the props and then, depending on which prop they actually collide with, different things may happen. The Player has a Rigidbody component and their collider is not set to be a trigger, but the colliders of the props are triggers. For now we’ll add some fields in the PlayerController class and use them in the OnTriggerEnter2D method. In most cases the prop the Player collides with should be destroyed, but the shield and the key should be collected and carried along.