Add timers and update them in the Update method
There will be a couple timers in the game. Let’s implement the timers that will ensure that the Player is slowed down, powered up, invisible or protected by shield after colliding with poison, a power drink, an invisibility cloak or a shield respectively only for a certain amount of time. The time will be checked and updated in the Update method. In the FixedUpdate method we’ll make sure that the Player moves at only half the speed when slowed down. We’ll also implement the RemoveShield to handle the shield when the time of protection is up.