Set up the monsters
We’ll set each monster’s features now. First we’ll add some variables to the MonsterController script and then we’ll set their values in the editor. In particular we’ll set the following features: speed, direction, walking distance and max health. We’ll also determine for each monster whether they’re mobile or not and whether they can shoot or not.