Add animators, animator controllers and animations
We’ll now create some animations for our characters and props. First we’ll add the Animator controller to each of them. Then we’ll add an animator controller to each animator. Finally, we’ll create all the animations for the Player, the Sorcerer, all the monsters and all the props.