Create the Door script to handle door opening and closing
Now if the Player has found the key, they should be able to open the door. In order to do that they must be close enough to the door. When, after opening the door, the Player moves farther away from it, the door should close. This functionality is in the Door script that we’ll attach to the Door prefab.