Implement the Damage method on the Player
If the Player bumps into something or gets shot (yet to be implemented), they take damage. Their current health decreases and they flash for a brief period of time. When the Player collides with a monster or a bomb, there’s a short knockback animation besides the flashing so that we can clearly see that a collision occurred.