Create the missile prefabs and assign them to the characters
Let’s now add some shooting functionality to the game. First we’ll create four missile prefabs (2 for the Player, 1 for the Sorcerer and 1 for the monsters) and then we’ll assign them to the characters. We’ll also create a Missile script and add it to the prefabs. Besides we’ll add a shoot point to each character. This is the transform empty GameObject where the missiles will be spawned.