Create the UI with placeholder data
We’ll create a canvas for our GUI now and save it as a prefab. It will contain four main parts: HUD (with Health image, Ammo text, Score text and High Score text with just some placeholder text for now), InfoUI (this will appear when a scene loads and display level info), PauseUI (a menu that will display when the game is paused), DoorText (info that will display near the door if the Player hasn’t found the key). Only the HUD will be set to active for now.