Create the GameManager prefab and add a GameManager script to it
We’ll now create a GameManager GameObject that will be used throughout the game to keep track of the Player’s data like score, high score and ammo. We’ll create a GameManager script that we will attach to the GameManager prefab. The script will take care of saving and loading data between scenes and between game sessions. It will also update the UI.