Add the InfoUI script to the InfoUI object
We want the level info to disappear after a while. While you can still see it, the game should be paused and as far as it’s gone, the game should resume. This will be handled by the InfoUI script. We’ll add methods for pausing and resuming the game to the GameManager script.