Implement effect sounds and modify the door script
Now every time a character shoots, collects something or gets hit, a sound effect should occur. Monsters and the Sorcerer are far away from the Player, so we’ll implement 3D sound for them. We’ll also modify the door script. Additionally we’ll add a left and right range GameObjects with a script to the door prefab to trigger the door opening and closing animations.